requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.3.0
autoflake==1.4
Automat==20.2.0
autopep8==1.5.3
bidict==0.21.2
black==20.8b1
cachetools==4.2.1
certifi==2020.4.5.2
cffi==1.14.5
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
distlib==0.3.0
eventlet==0.30.2
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.8
flask-heroku==0.1.9
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.7.0
Flask-OAuth==0.12
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.3
gevent==21.1.2
gunicorn==20.0.4
httplib2==0.18.1
hyperlink==21.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
libcst==0.3.16
limits==1.5.1
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.9
parsel==1.6.0
pathspec==0.8.0
Protego==0.1.16
protobuf==3.15.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyflakes==2.2.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-engineio==4.0.0
python-socketio==5.0.3
pytz==2020.5
PyYAML==5.3.1
queuelib==1.5.0
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
Twisted==21.2.0
typed-ast==1.4.1
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
w3lib==1.22.0
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.2.0
